Add fast swipe cmd
Added a faster swipe command that doesn't require a session.

Added a session-less pressButton map. It doesn't need a session to begin with...

Added a test keyboard events via synthetic events call. Tried it and it didn't seem to work but leaving the code in in case someone else is able to figure it out.