Change HTTP to NNG ; Rewritten logic for API Disabled HTTP API and added new NNG ( NanoMsg Next Gen ) API. The old HTTP server was problematic as it did not properly support reused HTTP connections, and so was exhausting file descriptors as many requests were made in a short period of time.
The logic within NNGServer of the new requests is new and uses as little of the existing WDA as possible.
Ultimately the goal is to rewrite the entirety of WDA trashing all the old logic and doing things in the cleanest most efficient ways possible.

Example is the new 'source' command, which uses the modern updated API call to Apple libraries. It actually outputs the correct frame positions unlike the existing one that outputs gibberish for those numbers that is inconsistent.

This is a breaking change. No existing WDA client will work as the HTTP server is disabled and will be deleted. An adapter will be needed for those wishing to use existing clients.

At first it was attempting to have HTTP and NNG side by side but this doesn't work because all XCUITest calls must happen from the main thread. An HTTP to NNG adapter may be added back later in its own thread to still enable both without an external adapter.

The new code files are not given with the same license. They are under the Anti-Corruption license, which has been included in the root of the project. Existing files are kept under the same license. You must, therefore, to use his project, accept both licenses.

NNG library has been committed as a prebuilt iOS library. This was built from an untouched clone of nanomsg/nng following the standard build instructions they provide here: https://github.com/nanomsg/nng/blob/master/docs/BUILD_IOS.adoc It is, therefore, under the MIT License.

Ujsonin library has been added as source. It remains under it's Anti-Corruption license. It has been modified to add a few features and fix some bugs.