Delete unused files and libraries
Deleted all of the unused routing and commands.
This does delete essentially all the old functionality, leaving only the commands in NNGServer and all the helper utilities and functions used by the old commands.

To be used in the same way as old WDA obviously many changes would be needed now.
This is a major divergence from the upstream, and it is likely that the new changes will become their own project instead of continuing to hack on this mishmash of old WDA code and newly written logic.

As it stands, many of the existing tests will/should now fail. Those should be removed also...

YYCache is no longer being used except to preserve a single fetched element to a later call. That will soon be changed; there is no intent to use any of the old caching logic as it is buggy and causes problems.