Moved to latest dumped Xcode Private Headers
Ran a new dump of Xcode Private XCTest headers using class-dump, and manually fixed up the dump to be able to run/function.

This replaces ALL of the old PrivateHeaders. Yay!

It made a lot of various issues more clear as to why things did/didn't work.

This change also removes the last remaining bits of WDA from the codebase!
CFA is now free of all tainting old Facebook junk.