Cleanup and add sysElPos command
The way systemApp was being handled did not work for interacting with elements of it. The systemApp cannot be cast directly to XCUIApplication and have it function in all ways due to restrictions in newer iOS / developer images.

Essentially when fetching the system app it must be treated like a snapshot provider only. You cannot get XCUIElement objects from it or searches like descendentsOf etc.